#1054 - support for `LTI_CONFIG_TEMPLATE_PATH`
To give admin users a way to replace the LTI configuration by URL template, added support for a `LTI_CONFIG_TEMPLATE_PATH` application setting.  It's normally null, which would default to the embedded LTI configuration template.  If not null, it must point to a valid file containing an LTI configuration template.