#1252 - upgrade Django to 3.1.7
After much research and testing, I decided to go with Django 3.1.7.  It appears to be the latest version available and it's one higher than highest version suggested by @dependabot.  I was cautiously thinking that we should only use the most recent version of 3.0.x, but MyLA seems to work well with 3.1.7.