#1252 - semi-related MySQL upgrade
While working on this issue, I found that MyLA's docker-compose.yml installed MySQL 5.7.22 although my DB was already configured to work with 5.7.26.  There are some changes between 22→26 that cause the older MySQL to complain about using a newer DB.  After discussion with @jonespm, we agreed it would probably be safe to specify the more generic 5.7.